- Add CancellationToken in clang scan
- Update test framework
- Fix compile failed

error CS1501: 方法 'InsertorUpdateImplementationLocationAsync' 沒有任何多載使用 3 個引數
- Add CancellationToken in Create project api
- Update logic of trigger cancel
- Lock project

+++ Arboris.EntityFramework/Repositories/ProjectRepository.cs
@@ -25,6 +25,8 @@ public class ProjectRepository(IDbContextFactory<ArborisDbContext> dbContextFact
         Project? project = await db.Projects.FindAsync(id);
         if (project is null)
             return Result.Fail("Project not found");
+        if (project.IsLocked)
+            return Result.Fail(new ExceptionalError(new InvalidOperationException("Project is locked")));
         db.Projects.Remove(project);
         await db.SaveChangesAsync();
         return Result.Ok();
@@ -34,7 +36,7 @@ public class ProjectRepository(IDbContextFactory<ArborisDbContext> dbContextFact
     {
         using ArborisDbContext db = await dbContextFactory.CreateDbContextAsync();
         Project[] oldProject = await db.Projects
-            .Where(item => item.CreateTime < DateTime.Now.AddDays(-days))
+            .Where(item => item.CreateTime < DateTime.Now.AddDays(-days) && !item.IsLocked)
             .ToArrayAsync();
         db.Projects.RemoveRange(oldProject);
         await db.SaveChangesAsync();
@@ -46,14 +48,26 @@ public class ProjectRepository(IDbContextFactory<ArborisDbContext> dbContextFact
         Project? project = await db.Projects.FindAsync(id);
         if (project is null)
             return Result.Fail<GetProject>("Project not found");
-        return Result.Ok(new GetProject(project.Id, project.CreateTime));
+        return Result.Ok(new GetProject(project.Id, project.SolutionName, project.CreateTime, project.IsLocked));
     }
 
     public async Task<GetProject[]> GetProjectsAsync()
     {
         using ArborisDbContext db = await dbContextFactory.CreateDbContextAsync();
         return await db.Projects
-            .Select(item => new GetProject(item.Id, item.CreateTime))
+            .Select(item => new GetProject(item.Id, item.SolutionName, item.CreateTime, item.IsLocked))
             .ToArrayAsync();
+    }
+
+    public async Task<Result> UpdateLockedAsync(Guid id, bool isLocked)
+    {
+        using ArborisDbContext db = await dbContextFactory.CreateDbContextAsync();
+        Project? project = await db.Projects.FindAsync(id);
+        if (project is null)
+            return Result.Fail("Project not found");
+        project.IsLocked = isLocked;
+        db.Projects.Update(project);
+        await db.SaveChangesAsync();
+        return Result.Ok();
     }
 }
